git-svn-id: http://financereport.googlecode.com/svn/FinancialReport@1168 d2ea8a1b-09b1-a710-b10d-4e27fa839f9d

+++ trunk/Java/DataLoadSpringAmar/src/com/pikefin/dao/inter/LogTweetsDao.java
@@ -12,6 +12,6 @@ public interface LogTweetsDao {
 	public Boolean deleteLogTweetsInfoById(Integer logTweetId ) throws GenericException;
 	public LogTweets loadLogTweetsInfo(Integer logTweetId) throws GenericException;
 	public List<LogTweets> loadAllLogTweets() throws GenericException;
-	public Integer getTweetCounts(Integer UserId) throws GenericException;
+	public Long getTweetCounts(Integer UserId) throws GenericException;
 
 }
