+++ trunk/Java/DataLoadSpringAmar/src/com/pikefin/services/inter/LogTweetsService.java
@@ -6,7 +6,7 @@ import com.pikefin.exceptions.GenericException;
 public interface LogTweetsService {
 
 	public LogTweets saveLogTweetsInfo(LogTweets logTweetsEntity) throws GenericException;
-	public Integer getTweetCounts(Integer UserId) throws GenericException;
+	public Long getTweetCounts(Integer UserId) throws GenericException;
 
 
 }
